requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.ini
#
amqp==5.3.1
    # via kombu
appdirs==1.4.4
    # via zeep
asgiref==3.8.1
    # via django
async-timeout==5.0.1
    # via redis
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
    #   service-identity
    #   twisted
    #   zeep
automat==24.8.1
    # via twisted
billiard==4.2.1
    # via celery
black==22.12.0
    # via -r requirements.ini
cached-property==2.0.1
    # via zeep
celery==5.4.0
    # via
    #   -r requirements.ini
    #   django-celery-beat
    #   flower
certifi==2024.12.14
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
constantly==23.10.4
    # via twisted
cron-descriptor==1.4.5
    # via django-celery-beat
cryptography==44.0.0
    # via
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.2.0
    # via
    #   parsel
    #   scrapy
defusedxml==0.7.1
    # via
    #   scrapy
    #   zeep
dj-database-url==2.3.0
    # via -r requirements.ini
django==4.2.18
    # via
    #   -r requirements.ini
    #   dj-database-url
    #   django-celery-beat
    #   django-cors-headers
    #   django-filter
    #   django-redis
    #   django-sonar
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   model-bakery
django-celery-beat==2.7.0
    # via -r requirements.ini
django-cors-headers==3.14.0
    # via -r requirements.ini
django-filter==24.3
    # via -r requirements.ini
django-redis==5.4.0
    # via -r requirements.ini
django-sonar==0.1.8
    # via -r requirements.ini
django-timezone-field==7.1
    # via django-celery-beat
djangorestframework==3.15.2
    # via
    #   -r requirements.ini
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.4.0
    # via -r requirements.ini
drf-spectacular==0.28.0
    # via -r requirements.ini
filelock==3.17.0
    # via tldextract
flower==2.0.1
    # via -r requirements.ini
gunicorn==20.1.0
    # via -r requirements.ini
humanize==4.11.0
    # via flower
hyperlink==21.0.0
    # via twisted
idna==3.10
    # via
    #   hyperlink
    #   requests
    #   tldextract
incremental==24.7.2
    # via twisted
inflection==0.5.1
    # via drf-spectacular
isodate==0.7.2
    # via zeep
itemadapter==0.10.0
    # via
    #   itemloaders
    #   scrapy
itemloaders==1.3.2
    # via scrapy
jmespath==1.0.1
    # via
    #   itemloaders
    #   parsel
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
kavenegar==1.1.2
    # via -r requirements.ini
kombu==5.4.2
    # via celery
lxml==5.3.0
    # via
    #   parsel
    #   scrapy
    #   zeep
model-bakery==1.11.0
    # via -r requirements.ini
mypy-extensions==1.0.0
    # via black
numpy==2.2.2
    # via pandas
packaging==24.2
    # via
    #   parsel
    #   scrapy
    #   scrapyd
pandas==2.2.3
    # via -r requirements.ini
parsel==1.10.0
    # via
    #   itemloaders
    #   scrapy
pathspec==0.12.1
    # via black
pillow==9.5.0
    # via -r requirements.ini
platformdirs==4.3.6
    # via black
prometheus-client==0.21.1
    # via flower
prompt-toolkit==3.0.50
    # via click-repl
protego==0.4.0
    # via scrapy
psutil==6.1.1
    # via -r requirements.ini
psycopg2-binary==2.9.10
    # via -r requirements.ini
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.1
    # via service-identity
pycparser==2.22
    # via cffi
pydispatcher==2.0.7
    # via scrapy
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pyopenssl==25.0.0
    # via scrapy
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   pandas
    #   python-crontab
python-decouple==3.8
    # via -r requirements.ini
python-dotenv==1.0.1
    # via -r requirements.ini
pytz==2024.2
    # via
    #   flower
    #   pandas
    #   zeep
pyyaml==6.0.2
    # via drf-spectacular
queuelib==1.7.0
    # via scrapy
redis==4.6.0
    # via
    #   -r requirements.ini
    #   django-redis
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   kavenegar
    #   requests-file
    #   requests-toolbelt
    #   tldextract
    #   zeep
requests-file==2.1.0
    # via tldextract
requests-toolbelt==1.0.0
    # via zeep
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
scrapy==2.12.0
    # via scrapyd
scrapyd==1.5.0
    # via -r requirements.ini
service-identity==24.2.0
    # via scrapy
six==1.17.0
    # via
    #   python-dateutil
    #   zeep
sqlparse==0.5.3
    # via
    #   django
    #   django-sonar
tldextract==5.1.3
    # via scrapy
tomli==2.2.1
    # via
    #   black
    #   incremental
tornado==6.4.2
    # via flower
twisted==24.11.0
    # via
    #   scrapy
    #   scrapyd
typing-extensions==4.12.2
    # via
    #   asgiref
    #   dj-database-url
    #   pyopenssl
    #   referencing
    #   twisted
tzdata==2025.1
    # via
    #   celery
    #   django-celery-beat
    #   kombu
    #   pandas
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.3.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
w3lib==2.3.1
    # via
    #   parsel
    #   scrapy
    #   scrapyd
wcwidth==0.2.13
    # via prompt-toolkit
zeep==3.4.0
    # via -r requirements.ini
zope-interface==7.2
    # via
    #   scrapy
    #   scrapyd
    #   twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
